requirements.txt
numpy
scipy
pandas
matplotlib
seaborn
tqdm
opencv-python
scikit-learn
scikit-image
tslearn
tensorboardX
torch==1.11.0+cu115
torchvision==0.12.0+cu115
pytorch-lightning
pycocotools
Cython
easydict==1.9
PyYAML
json-tricks
yacs
cffi
munkres
shapely
motmetrics
colorama
loguru
thop
tabulate
xlrd
openpyxl
optuna
optuna-dashboard
